Fixed key hooks
Still there is a small issue. When i dont put my character close to a wall or onto a corner and if i press a key from wasd my character stays where it is and a new character shows up and it moves. But if i come onto my old player picture it disappears and now there becomes one character available and continues moving when i press the keys