fixed exit issue.
It was only exit the exit door when you collect all of the collectibles and reached the exit door but if i didnt collect collectibles and reached the exit door, it would disappear because of if (data->map[y][x] == 'E' && data->collectible == 0) this function but i separated my move_player function and added a helper function as exit_door. I gave this if (data->map[y][x] == 'E'){if (data->collectible > 0) ...} so that meant if there is still any collectible left you cant go through the exit and with that the exit door didn't disappear.